[mod]CTP行情接口的成交量类型为int，但是部分品种，如ETF现货以及指数，成交量都超过int的范围，导致显示数据为负数，所以成交量转换的地方，加了一个强制转成uint32_t的逻辑，不然拿到的就是一个负数

+++ src/ParserCTP/ParserCTP.cpp
@@ -285,7 +285,7 @@ void ParserCTP::OnRtnDepthMarketData( CThostFtdcDepthMarketDataField *pDepthMark
 	quote.open = checkValid(pDepthMarketData->OpenPrice);
 	quote.high = checkValid(pDepthMarketData->HighestPrice);
 	quote.low = checkValid(pDepthMarketData->LowestPrice);
-	quote.total_volume = pDepthMarketData->Volume;
+	quote.total_volume = (uint32_t)pDepthMarketData->Volume;
 	quote.trading_date = m_uTradingDate;
 	if(pDepthMarketData->SettlementPrice != DBL_MAX)
 		quote.settle_price = checkValid(pDepthMarketData->SettlementPrice);
